test(e2e): group all transformers tests into one singe place
