+++ e2e/__cases__/ast-transformers/hoisting/__test_modules__/a.ts
