+++ e2e/__cases__/ast-transformers/hoisting/general-hoisting.spec.ts
