+++ e2e/__cases__/ast-transformers/path-mapping/direct-import.spec.ts
