+++ e2e/__cases__/ast-transformers/path-mapping/dynamic-import.spec.ts
