+++ e2e/__cases__/ast-transformers/path-mapping/import-default.spec.ts
