+++ e2e/__cases__/ast-transformers/path-mapping/import-legacy.spec.ts
