+++ e2e/__cases__/ast-transformers/path-mapping/import-require.spec.ts
