+++ e2e/__cases__/ast-transformers/path-mapping/share/export.ts
