+++ e2e/__cases__/ast-transformers/path-mapping/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "baseUrl": ".",
+    "baseUrl": "./",
     "paths": {
       "@share/*": ["share/*"]
     }
