--- e2e/__tests__/hoisting.test.ts
@@ -1,20 +1,0 @@
-import { allValidPackageSets } from '../__helpers__/templates'
-import { configureTestCase } from '../__helpers__/test-case'
-
-describe('Hoisting', () => {
-  const testCase = configureTestCase('hoisting', {
-    writeIo: true,
-    jestConfig: {
-      testEnvironment: 'node',
-      automock: true,
-    },
-  })
-
-  testCase.runWithTemplates(allValidPackageSets, 0, (runTest, { testLabel }) => {
-    it(testLabel, () => {
-      const result = runTest()
-      expect(result.status).toBe(0)
-      expect(result).toMatchSnapshot(testLabel)
-    })
-  })
-})
